+++ lib/presentation/credentials/credentials_controller.dart
@@ -7,44 +7,48 @@ class CredentialsController extends BaseController {
   final CredentialsStore _credentialsStore;
   final ListCredentialUseCase _listCredentialUseCase;
   final UpdateCredentialUseCase _updateCredentialUseCase;
+  final CredentialRepository _repository;
 
   CredentialsController({
     required CredentialsStore credentialsStore,
     required ListCredentialUseCase listCredentialUseCase,
     required UpdateCredentialUseCase updateCredentialUseCase,
+    required CredentialRepository credentialRepository,
   }) : _credentialsStore = credentialsStore,
        _listCredentialUseCase = listCredentialUseCase,
-       _updateCredentialUseCase = updateCredentialUseCase;
+       _updateCredentialUseCase = updateCredentialUseCase,
+       _repository = credentialRepository;
 
-  RxList<CredentialEntity> get credentials => _credentialsStore.credentials;
-  RxBool isLoading = RxBool(false);
-  RxString errorMessage = RxString('');
+  ValueNotifier<List<ValueNotifier<CredentialEntity>>> get credentials =>
+      _repository.valueListenable;
+  ValueNotifier<bool> isLoading = ValueNotifier<bool>(true);
+  ValueNotifier<String> errorMessage = ValueNotifier<String>('');
 
   ScrollController? scrollController;
 
   @override
-  void onReady() {
+  Future<void> onReady() async {
     super.onReady();
+    await _repository.initLocalDatabase();
     getAllCredentials();
   }
 
   Future<void> getAllCredentials() async {
     final track = CrashlyticsServiceManager.instance.trackOperation(
       name: 'get-all-credentials-performance-tracking',
-      operation: 'get-all-credentials',
     );
     try {
       isLoading.value = true;
       errorMessage.value = '';
-      credentials.value = await _listCredentialUseCase.call();
+      await _listCredentialUseCase.call();
     } on BaseException catch (error) {
-      Log.error('getAllCredentials', error: error);
+      Log.error(error, StackTrace.current);
       errorMessage.value = error.message.tr;
-      track.catchError(error: error);
+      track.setStatus(TrackOperationStatus.internalError);
     } catch (error, stackTrace) {
-      Log.error('getAllCredentials', error: error, stackTrace: stackTrace);
+      Log.error(error, stackTrace);
       errorMessage.value = error.toString();
-      track.catchError(error: error);
+      track.setStatus(TrackOperationStatus.internalError);
     } finally {
       isLoading.value = false;
       track.finish();
@@ -54,7 +58,6 @@ class CredentialsController extends BaseController {
   Future<void> addCredential() async {
     _credentialsStore.credential.value = null;
     final int? scrollToIndex = await AppNavigator.toNamed(AppRouter.credential);
-    _listCredentialsToSaveLocally();
     if (scrollToIndex != null) {
       scrollController?.jumpTo(scrollToIndex * 56);
     }
@@ -63,15 +66,27 @@ class CredentialsController extends BaseController {
   Future<void> openCredential(CredentialEntity item) async {
     _credentialsStore.credential.value = item;
     final int? scrollToIndex = await AppNavigator.toNamed(AppRouter.credential);
-    _listCredentialsToSaveLocally();
     if (scrollToIndex != null) {
       scrollController?.jumpTo(scrollToIndex * 56);
     }
   }
 
-  void _listCredentialsToSaveLocally() {
+  Future<void> errorFavIcon(CredentialEntity item) async {
     try {
-      _listCredentialUseCase.call();
+      final tempCredential = CredentialEntity(
+        objectId: item.objectId,
+        name: item.name,
+        userName: item.userName,
+        password: item.password,
+        secretKeyOTP: item.secretKeyOTP,
+        url: item.url,
+        faviconUrl: null,
+        notes: item.notes,
+        createdAt: item.createdAt,
+        updatedAt: item.updatedAt,
+      );
+
+      await _updateCredentialUseCase.call(tempCredential);
     } catch (_) {}
   }
 }